test: relocate new tests into vitest unit/nuxt projects so npm test collects them
